requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.6
django-rest-framework==0.1.0
djangorestframework==3.15.1
exceptiongroup==1.2.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
openai==0.28.0
pdf2image==1.17.0
pillow==10.3.0
pydantic==2.7.2
pydantic_core==2.18.3
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
yarl==1.9.4
